Mapped F10 to Zed Menu and F1 to Help
Per the CUA guidelines, F10 should bring up the main menu and F1 should
bring up help. I need to see if I can find a bit more detail about the
behavior of F10 and F1 to ensure my fix is correct but this is a step in
the right direction.